Add TransferDomain contract and contract cleanup
