+++ lib/ain-contracts/Cargo.toml
@@ -11,7 +11,6 @@ anyhow = "1.0"
 ethereum = "0.14.0"
 serde_json = "1.0.96"
 hex = "0.4.3"
-lazy_static = "1.4"
 sha3 = "0.10.6"
 sp-core = "20.0.0"
