+++ lib/ain-evm/src/evm.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use ain_contracts::{Contracts, CONTRACT_ADDRESSES};
+use ain_contracts::{Contract, DST20Contract, ReservedContract};
 use anyhow::format_err;
 use ethereum::{
     Block, EIP1559ReceiptData, LegacyTransaction, PartialHeader, ReceiptV3, TransactionAction,
@@ -13,14 +13,13 @@ use primitive_types::H256;
 
 use crate::backend::{EVMBackend, Vicinity};
 use crate::block::BlockService;
-use crate::bytes::Bytes;
+use crate::contract::{self, DST20BridgeInfo, DeployContractInfo};
 use crate::core::{EVMCoreService, XHash};
 use crate::executor::{AinExecutor, TxResponse};
 use crate::fee::{calculate_gas_fee, calculate_prepay_gas_fee};
 use crate::filters::FilterService;
 use crate::log::LogService;
 use crate::receipt::ReceiptService;
-use crate::services::SERVICES;
 use crate::storage::traits::BlockStorage;
 use crate::storage::Storage;
 use crate::traits::Executor;
@@ -45,17 +44,6 @@ pub struct FinalizedBlockInfo {
     pub total_burnt_fees: U256,
     pub total_priority_fees: U256,
     pub block_number: U256,
-}
-
-pub struct DeployContractInfo {
-    pub address: H160,
-    pub storage: Vec<(H256, H256)>,
-    pub bytecode: Bytes,
-}
-
-pub struct DST20BridgeInfo {
-    pub address: H160,
-    pub storage: Vec<(H256, H256)>,
 }
 
 pub type ReceiptAndOptionalContractAddress = (ReceiptV3, Option<H160>);
@@ -201,13 +189,30 @@ impl EVMServices {
                 address,
                 storage,
                 bytecode,
-            } = EVMServices::counter_contract(dvm_block_number, current_block_number)?;
+            } = contract::counter_contract(
+                &executor.backend,
+                dvm_block_number,
+                current_block_number,
+            )?;
+            executor.deploy_contract(address, bytecode, storage)?;
+
+            // Deploy transfer domain contract on the first block
+            let DeployContractInfo {
+                address,
+                storage,
+                bytecode,
+            } = contract::transfer_domain_contract()?;
             executor.deploy_contract(address, bytecode, storage)?;
         } else {
             // Ensure that state root changes by updating counter contract storage
             let DeployContractInfo {
                 address, storage, ..
-            } = EVMServices::counter_contract(dvm_block_number, current_block_number)?;
+            } = contract::counter_contract(
+                &executor.backend,
+                dvm_block_number,
+                current_block_number,
+            )?;
+            debug!("address.as_fixed_bytes() : {:#?}", address.as_fixed_bytes());
             executor.update_storage(address, storage)?;
         }
 
@@ -258,6 +263,7 @@ impl EVMServices {
                         "[construct_block] Transfer domain to EVM for address {:x?}, amount: {}, queue_id {}, tx hash {}",
                         to, signed_tx.value(), queue_id, queue_item.tx_hash
                     );
+
                     if let Err(e) = executor.add_balance(to, signed_tx.value()) {
                         debug!("[construct_block] EvmIn failed with {e}");
                         failed_transactions.push(queue_item.tx_hash);
@@ -297,7 +303,7 @@ impl EVMServices {
                         address,
                         bytecode,
                         storage,
-                    } = EVMServices::dst20_contract(&mut executor, address, name, symbol)?;
+                    } = contract::dst20_contract(&executor.backend, address, name, symbol)?;
 
                     if let Err(e) = executor.deploy_contract(address, bytecode.clone(), storage) {
                         debug!("[construct_block] EvmOut failed with {e}");
@@ -318,7 +324,7 @@ impl EVMServices {
                         to, contract, amount, out
                     );
 
-                    match EVMServices::bridge_dst20(&mut executor, contract, to, amount, out) {
+                    match contract::bridge_dst20(&executor.backend, contract, to, amount, out) {
                         Ok(DST20BridgeInfo { address, storage }) => {
                             if let Err(e) = executor.update_storage(address, storage) {
                                 debug!("[construct_block] EvmOut failed with {e}");
@@ -474,129 +480,6 @@ impl EVMServices {
         Ok(())
     }
 
-    /// Returns address, bytecode and storage with incremented count for the counter contract
-    pub fn counter_contract(
-        dvm_block_number: u64,
-        evm_block_number: U256,
-    ) -> Result<DeployContractInfo> {
-        let address = *CONTRACT_ADDRESSES.get(&Contracts::CounterContract).unwrap();
-        let bytecode = ain_contracts::get_counter_bytecode()?;
-        let count = SERVICES
-            .evm
-            .core
-            .get_latest_contract_storage(address, ain_contracts::u256_to_h256(U256::one()))?;
-
-        debug!("Count: {:#x}", count + U256::one());
-
-        Ok(DeployContractInfo {
-            address,
-            bytecode: Bytes::from(bytecode),
-            storage: vec![
-                (
-                    H256::from_low_u64_be(0),
-                    ain_contracts::u256_to_h256(U256::one()),
-                ),
-                (
-                    H256::from_low_u64_be(1),
-                    ain_contracts::u256_to_h256(evm_block_number),
-                ),
-                (
-                    H256::from_low_u64_be(2),
-                    ain_contracts::u256_to_h256(U256::from(dvm_block_number)),
-                ),
-            ],
-        })
-    }
-
-    pub fn dst20_contract(
-        executor: &mut AinExecutor,
-        address: H160,
-        name: String,
-        symbol: String,
-    ) -> Result<DeployContractInfo> {
-        match executor.backend.get_account(&address) {
-            None => {}
-            Some(account) => {
-                if account.code_hash != ain_contracts::get_system_reserved_codehash()? {
-                    return Err(format_err!("Token address is already in use").into());
-                }
-            }
-        }
-
-        let bytecode = ain_contracts::get_dst20_bytecode()?;
-        let storage = vec![
-            (
-                H256::from_low_u64_be(3),
-                ain_contracts::get_abi_encoded_string(name.as_str()),
-            ),
-            (
-                H256::from_low_u64_be(4),
-                ain_contracts::get_abi_encoded_string(symbol.as_str()),
-            ),
-        ];
-
-        Ok(DeployContractInfo {
-            address,
-            bytecode: Bytes::from(bytecode),
-            storage,
-        })
-    }
-
-    pub fn bridge_dst20(
-        executor: &mut AinExecutor,
-        contract: H160,
-        to: H160,
-        amount: U256,
-        out: bool,
-    ) -> Result<DST20BridgeInfo> {
-        // check if code of address matches DST20 bytecode
-        let account = executor
-            .backend
-            .get_account(&contract)
-            .ok_or_else(|| format_err!("DST20 token address is not a contract"))?;
-
-        if account.code_hash != ain_contracts::get_dst20_codehash()? {
-            return Err(format_err!("DST20 token code is not valid").into());
-        }
-
-        let storage_index = ain_contracts::get_address_storage_index(to);
-        let balance = executor
-            .backend
-            .get_contract_storage(contract, storage_index.as_bytes())?;
-
-        let total_supply_index = H256::from_low_u64_be(2);
-        let total_supply = executor
-            .backend
-            .get_contract_storage(contract, total_supply_index.as_bytes())?;
-
-        let (new_balance, new_total_supply) = if out {
-            (
-                balance.checked_sub(amount),
-                total_supply.checked_sub(amount),
-            )
-        } else {
-            (
-                balance.checked_add(amount),
-                total_supply.checked_add(amount),
-            )
-        };
-
-        let new_balance = new_balance.ok_or_else(|| format_err!("Balance overflow/underflow"))?;
-        let new_total_supply =
-            new_total_supply.ok_or_else(|| format_err!("Total supply overflow/underflow"))?;
-
-        Ok(DST20BridgeInfo {
-            address: contract,
-            storage: vec![
-                (storage_index, ain_contracts::u256_to_h256(new_balance)),
-                (
-                    total_supply_index,
-                    ain_contracts::u256_to_h256(new_total_supply),
-                ),
-            ],
-        })
-    }
-
     ///
     /// # Safety
     ///
@@ -621,7 +504,7 @@ impl EVMServices {
         match backend.get_account(&address) {
             None => {}
             Some(account) => {
-                if account.code_hash == ain_contracts::get_dst20_codehash()? {
+                if account.code_hash == DST20Contract::codehash()? {
                     return Ok(true);
                 }
             }
@@ -640,7 +523,7 @@ impl EVMServices {
     }
 
     pub fn reserve_dst20_namespace(&self, executor: &mut AinExecutor) -> Result<()> {
-        let bytecode = ain_contracts::get_system_reserved_bytecode()?;
+        let bytecode = ReservedContract::bytecode()?;
         let addresses = (1..=1024)
             .map(|token_id| ain_contracts::dst20_address_from_token_id(token_id).unwrap())
             .collect::<Vec<H160>>();
